+++ modules/afk_mover/configuration.example.py
@@ -1,0 +1,3 @@
+AFK_MOVED_MESSAGE = 'You were moved because of inactivity.'
+BUSY_CHANNEL = 345  # Busy
+CHECK_AFK_INTERVAL_SECONDS = 10
