+++ modules/afk_mover/move_afk.py
@@ -1,9 +1,9 @@
 import time
-from TS3Bot import TS3Bot
-from configuration import BUSY_CHANNEL, CHECK_AFK_INTERVAL_SECONDS, AFK_MOVED_MESSAGE
+from ts3bot import TS3Bot
+from .configuration import BUSY_CHANNEL, CHECK_AFK_INTERVAL_SECONDS, AFK_MOVED_MESSAGE
 
 
-def move_afk(bot: TS3Bot):
+def afk_mover(bot: TS3Bot):
     while 1:
         for client in bot.create_client_list():
             if client.is_afk and not client.in_afk_channel:
