+++ modules/doodle/configuration.example.py
@@ -1,0 +1,12 @@
+BANNERS = {
+    'default': 'https://your-domain.com/files/banner-default.png',
+    'covid': 'https://your-domain.com/files/banner-corona.png',
+    'valentines': 'https://your-domain.com/files/banner-valentines.png',
+    'eastern': 'https://your-domain.com/files/banner-eastern.png',
+    'april fools': 'https://your-domain.com/files/banner-april-fools.png',
+    'mothers day': 'https://your-domain.com/files/banner-mothers-day.png',
+    'fathers day': 'https://your-domain.com/files/banner-fathers-day.png',
+    'halloween': 'https://your-domain.com/files/banner-halloween.png',
+    'christmas': 'https://your-domain.com/files/banner-christmas.png',
+    'new years': 'https://your-domain.com/files/banner-new-year.png',
+}
