+++ modules/games/configuration.example.py
@@ -1,0 +1,23 @@
+ACCOUNTS_DB_NAME = 'accounts'
+ACCOUNT_START_BALANCE = 5000
+ACCOUNT_INFO_BALANCE = 'Your balance is {}.'
+
+GAME_INVALID_CMD = 'Please enter a valid command.'
+GAME_INVALID_AMOUNT = 'Please enter a valid amount.'
+GAME_INVALID_GAME = 'Please enter a valid game.'
+GAME_ABORT_BAD_BET = 'Please enter a valid bet!'
+GAME_ABORT_BROKE = 'Sorry, you don\'t have enough money!'
+GAME_ABORT_BAD_WAGER = 'Sorry, minimum wager is {}!'
+GAME_GREETINGS = 'Hello {}. Thank you for playing!'  # player_name
+GAME_WON_MSG = 'You win [b][color={0}]{1}[/color][/b]!!'  # color, win_amount
+GAME_LOST_MSG = 'Oh no! You lost [b][color=firebrick]{}[/color][/b]!'  # wager
+
+# Slots
+SLOTS_MIN_WAGER = 50
+SLOTS_START_MSG = 'You started the slot machine! Good luck!'
+SLOTS_RESULT_MSG = 'Your reels show:'
+
+# Roulette
+ROULETTE_MIN_WAGER = 50
+ROULETTE_START_MSG = 'You bet [b]{0}[/b] on {1}! Let\'s spin the wheel!'  # amount, number or name
+ROULETTE_RESULT_MSG = 'The ball lands on [b]{}[/b]!!'  # winning_number
