+++ modules/reminder/configuration.example.py
@@ -1,0 +1,2 @@
+REMINDER_3_DAYS_MSG = 'Crackerbarrel meeting in 3 days!'
+REMINDER_TODAY_MSG = 'Crackerbarrel meeting today 8pm!'
