+++ modules/reminder/reminder.py
@@ -1,8 +1,8 @@
 import calendar
 import datetime
 
-from TS3Bot import TS3Bot
-from configuration import CRACKERBARREL_REMINDER_3_DAYS, CRACKERBARREL_REMINDER_TODAY
+from ts3bot import TS3Bot
+from .configuration import REMINDER_3_DAYS_MSG, REMINDER_TODAY_MSG
 
 days = {
     'Monday': 0,
@@ -22,10 +22,10 @@ def crackerbarrel_reminder(bot: TS3Bot):
 
     if today.day == last_wednesday:
         bot.set_host_message_mode(2)
-        return bot.set_new_host_message(CRACKERBARREL_REMINDER_TODAY)
+        return bot.set_new_host_message(REMINDER_TODAY_MSG)
     elif today.day == last_wednesday - 3:
         bot.set_host_message_mode(2)
-        return bot.set_new_host_message(CRACKERBARREL_REMINDER_3_DAYS)
+        return bot.set_new_host_message(REMINDER_3_DAYS_MSG)
     elif today.day == last_wednesday + 1:
         return bot.reset_host_message()
