+++ modules/time_tracker/configuration.example.py
@@ -1,0 +1,11 @@
+# Tracking
+PROFILES_DB_NAME = 'profiles'
+MEASUREMENT_INTERVAL_SECONDS = 5
+MEASUREMENT_EMPTY_INTERVAL_SECONDS = 5 * 60
+TRACK_TOGGLE_CMD = '!toggle'
+TRACKING_INFO_MSG = f'Your connection time is being tracked. To opt-out, enter [b]{TRACK_TOGGLE_CMD}[/b].'
+DO_NOT_TRACK_CONFIRMED_MSG = f'Your connection time is not being tracked anymore. ' \
+                             f'[b]{TRACK_TOGGLE_CMD}[/b] to undo.'
+
+# WordPress
+WORDPRESS_UPDATE_INTERVAL_SECONDS = 15 * 60
