+++ modules/time_tracker/time_tracker.py
@@ -1,13 +1,13 @@
 import time
 
-from configuration import TRACK_TOGGLE_CMD, TRACKING_INFO_MSG, DO_NOT_TRACK_CONFIRMED_MSG, \
+from .configuration import TRACK_TOGGLE_CMD, TRACKING_INFO_MSG, DO_NOT_TRACK_CONFIRMED_MSG, \
     MEASUREMENT_EMPTY_INTERVAL_SECONDS, MEASUREMENT_INTERVAL_SECONDS
 
-from modules.time_tracker.ProfilesDB import ProfilesDB
-from TS3Bot import TS3Bot
+from modules.time_tracker.profiles_db import ProfilesDB
+from ts3bot import TS3Bot
 
 
-def start(bot: TS3Bot, database: ProfilesDB):
+def start_tracker(bot: TS3Bot, database: ProfilesDB):
     messaged_clients = list()
     first_timecheck = round(time.time(), 3)
     while 1:
@@ -28,10 +28,10 @@ def start(bot: TS3Bot, database: ProfilesDB):
                 continue
 
             do_not_track = database.get_do_not_track(client.uid)
-            __inform_client(bot, client.id, do_not_track)
+            _inform_client(bot, client.id, do_not_track)
             messaged_clients.append(client.b64_uid)
 
-        __check_toggle_messages(bot, database)
+        _check_toggle_messages(bot, database)
 
         time.sleep(MEASUREMENT_INTERVAL_SECONDS)
 
@@ -63,7 +63,7 @@ def start(bot: TS3Bot, database: ProfilesDB):
             database.update_profile_afk(client, new_afk)
 
 
-def __check_toggle_messages(bot, database):
+def _check_toggle_messages(bot, database):
     for message in bot.unused_messages:
         if message.content != TRACK_TOGGLE_CMD:
             continue
@@ -71,10 +71,10 @@ def __check_toggle_messages(bot, database):
         message.mark_as_used()
 
         do_not_track = database.toggle_do_not_track(message.invoker_uid)
-        __inform_client(bot, message.invoker_id, do_not_track)
+        _inform_client(bot, message.invoker_id, do_not_track)
 
 
-def __inform_client(bot, client_id: int, do_not_track: bool):
+def _inform_client(bot, client_id: int, do_not_track: bool):
     if do_not_track:
         msg = DO_NOT_TRACK_CONFIRMED_MSG
     else:
